Restructure Project (frontend)
