+++ frontend/src/components/dashboard/LiveWebSocket.tsx
